requirements.txt
# ── Core LangChain (pinned to 0.2.x) ────────────────────────────
langchain>=0.2.0,<0.3.0
langchain-core>=0.2.0,<0.3.0
langchain-community>=0.2.0,<0.3.0
langchain-text-splitters>=0.2.0,<0.3.0

# ── Vector store ─────────────────────────────────────────────────
faiss-cpu

# ── Embeddings ───────────────────────────────────────────────────
sentence-transformers

# ── PDF parsing ──────────────────────────────────────────────────
pypdf

# ── API server ───────────────────────────────────────────────────
fastapi
uvicorn[standard]

# ── Utilities ────────────────────────────────────────────────────
python-dotenv

# ── LLM API providers (ChatOpenAI wrapper — works with Groq/Together too)
openai
